Version 3.3
Switched to optimized library for ILI9341 (which also works for ILI9340)
from PJRC. Screen refresh rate is now 12.5ms!! No longer suppressing
screen update while logging data. Add bar graphs to the right of each
value. This increased the refresh rate to 25-30ms (depending on values).
However, if overclocked to 120Mhz it drops to 20-25ms. Had to remove the
cochlea bitmap to make room for the graphs.