Version 3.4
Moving to battery powered handheld device.  NOTE: in order to program
via USB, the power button must be in the 'on' position! I believe this
is due to the ground wire of the switch being connected when the switch
is 'off'. Added analog input to monitor battery voltage. Using correct
values of resistors in the voltage divider gives a value about 9mV
higher than battery voltage. Added biasing button. If bias performed
before logging is started, bias values are stored as first line in the
csv file. Otherwise, it will be evident in the file because all values
will simultaneously go to zero (and the bias values will be the average
of the previous 100 samples). Added restart button. Updated the
ILI9341_t3 library that included a fix for the omitting of the bottom
part of characters such as g and y.